+++ src/api/auth/user/handleKakaoLogin.js
@@ -1,33 +1,29 @@
-import { API_URL, KAKAO_LOGIN } from "./config";
+import { API_URL, KAKAO_LOGIN } from "../../config.js";
 
-export const requestKakaoLogin = () => {
+export const handleKakaoLogin = () => {
   return new Promise((resolve) => {
     window.open(KAKAO_LOGIN, "resizable=no,location=no,scrollbars=yes");
 
     const messageHandler = (event) => {
-      const allowedOrigins = [
-        API_URL,
-        "http://192.168.35.254:5173",
-        "http://localhost:5173",
-      ];
+      const allowedOrigins = [API_URL, "http://localhost:5173"];
       if (allowedOrigins.includes(event.origin) === false) {
-        console.error("[requestKakaoLogin] Invalid origin: ", event.origin);
+        console.error("[handleKakaoLogin] Invalid origin: ", event.origin);
         return; // 서버 주소 확인
       }
-      console.log("[requestKakaoLogin] event.data:", event.data);
-      console.log("[requestKakaoLogin] event.origin:", event.origin);
+      console.log("[handleKakaoLogin] event.data:", event.data);
+      console.log("[handleKakaoLogin] event.origin:", event.origin);
 
       // 사용자가 등록되지 않은 경우
       if ("not_registered_user" in event.data) {
         const { email } = event.data.not_registered_user;
         console.log(
-          "[requestKakaoLogin] User not found, please redirect to register page"
+          "[handleKakaoLogin] User not found, please redirect to register page"
         );
         resolve({ status: false, user: { email } }); // Promise 해결
         // setTimeout(() => {
         //   window.location.reload();
         //   console.log(
-        //     "[requestKakaoLogin] 창이 닫히고, 새로고침을 예약했습니다."
+        //     "[handleKakaoLogin] 창이 닫히고, 새로고침을 예약했습니다."
         //   );
         // }, 500);
       } else {
@@ -36,7 +32,7 @@ export const requestKakaoLogin = () => {
         // setTimeout(() => {
         //   window.location.reload();
         //   console.log(
-        //     "[requestKakaoLogin] 창이 닫히고, 새로고침을 예약했습니다."
+        //     "[handleKakaoLogin] 창이 닫히고, 새로고침을 예약했습니다."
         //   );
         // }, 500);
       }
