+++ src/api/conversations/chat/getMessageInChatroomById.js
@@ -1,13 +1,13 @@
 import meet from "./checkAuthorized.js";
-import { GET_CHAT_BY_CHATROOM_ID } from "./config";
+import { GET_CHAT_BY_CHATROOM_ID } from "../../config.js";
 
-export const getChatByChatroomId = async (chatroomId) => {
+export const getMessageInChatroomById = async (chatroomId) => {
   try {
     const response = await meet.get(GET_CHAT_BY_CHATROOM_ID, {
       params: { chatroom_id: chatroomId },
     });
 
-    console.log("[getChatByChatroomId] GET chat:", response.data);
+    console.log("[getMessageInChatroomById] GET chat:", response.data);
     return response.data;
   } catch (error) {
     console.error(
