+++ src/hooks/useUserProfile.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
-import { getUserProfile } from "../api/getUserProfile";
-import { reissueToken } from "../api/reissueToken";
+import { getUserProfile } from "../api/auth/user/getUserProfile.js";
+import { reissueToken } from "../api/auth/user/reissueToken.js";
 
 const useUserProfile = () => {
   const [userProfile, setUserProfile] = useState(null);
